added cylinder. found/fixed(?) problem in materials with cos terms.
I added flipNormal() earlier to get the normal on the same side as
the incomingDir so reflect()/refract()/cosSampleHemisphere() etc
would work right. Elsewhere I was calculating cos terms using
dot(n,wo). If the normal and wo were pointing on differt 'sides',
the cos term would be negative, and causing subtle (and some not
subtle) errors in the calculated colors.

I finally found the error because the inside of the cylinder was
not working right, and it seemed like a problem with the normal.
In a way it was, but really the problem was the with negative cos
terms in the LTE and elsewhere, the final pixel color was becoming
negative, and therefore clamped to 0 (black). The solution was to
take the abs() of cos terms.